// frontend/src/types/schedule.ts
export type EmployeeTimeOff = {
    id: number;
    employeeId: number;
    employeeName: string;
    startDate: Date;
    endDate: Date;
    type: TimeOffType;
}

export enum TimeOffType {
    Vacation = "Vacation",
    SickLeave = "SickLeave",
    PersonalDay = "PersonalDay",
}